+++ tests/integration/test_containers.py
@@ -15,7 +15,7 @@ if system() == "Darwin":
         "test_cbz.cbz": (93408, 84493, ("cbz", 84493)),
         "test_cbr.cbr": (93725, 93725, ("cbz", 84506)),
         "test_rar.rar": (93675, 93675, ("zip", 84493)),
-        "test_zip.zip": (7783, 5092, ("zip", 5092)),
+        "test_zip.zip": (7783, 6918, ("zip", 6918)),
         "igp-twss.epub": (292448, 281248, ("epub", 281248)),
     }
 else:
@@ -23,7 +23,7 @@ else:
         "test_cbz.cbz": (93408, 84481, ("cbz", 84481)),
         "test_cbr.cbr": (93725, 93725, ("cbz", 84494)),
         "test_rar.rar": (93675, 93675, ("zip", 84481)),
-        "test_zip.zip": (7783, 5092, ("zip", 5092)),
+        "test_zip.zip": (7783, 6918, ("zip", 6918)),
         "igp-twss.epub": (292448, 280700, ("epub", 280700)),
     }