+++ tests/integration/test_images_dir.py
@@ -12,9 +12,9 @@ TMP_ROOT = get_test_dir()
 FNS = {
     "07themecamplist.pdf": (93676, 93676, ("pdf", 93676), ("pdf", 93676)),
     "test_animated_gif.gif": (16383, 16358, ("png", 13411), ("webp", 11866)),
-    "test_animated_png.png": (63435, 63435, ("png", 63435), ("webp", 41694)),
+    "test_animated_png.png": (63435, 63435, ("png", 63435), ("webp", 53522)),
     "test_animated_webp.webp": (13610, 13610, ("webp", 13610), ("webp", 13610)),
-    "test_png.png": (7967, 4149, ("png", 4149), ("webp", 2044)),
+    "test_png.png": (7967, 4149, ("png", 4149), ("webp", 3870)),
     "test_pre-optimized_jpg.jpg": (
         22664,
         22664,
@@ -23,41 +23,41 @@ FNS = {
     ),
     "test_svg.svg": (5393, 4871, ("svg", 4871), ("svg", 4871)),
     "test_txt.txt": (6, 6, ("txt", 6), ("txt", 6)),
-    "test_webp_lossless.webp": (5334, 2044, ("webp", 2044), ("webp", 2044)),
+    "test_webp_lossless.webp": (5334, 3870, ("webp", 3870), ("webp", 3870)),
     "test_webp_lossless_pre-optimized.webp": (
         3798,
         3798,
         ("webp", 3798),
         ("webp", 3798),
     ),
-    "mri.tif": (230578, 230578, ("png", 131923), ("webp", 73242)),
+    "mri.tif": (230578, 230578, ("png", 131923), ("webp", 114740)),
     "test_webp_lossy.webp": (2764, 2764, ("webp", 2764), ("webp", 2764)),
 }
 if platform.system() == "Darwin":
     FNS.update(
         {
-            "test_bmp.bmp": (141430, 141430, ("png", 67215), ("webp", 30530)),
+            "test_bmp.bmp": (141430, 141430, ("png", 67215), ("webp", 47524)),
             "test_gif.gif": (138952, 138944, ("png", 112137), ("webp", 107924)),
             "test_jpg.jpg": (97373, 87913, ("jpg", 87913), ("jpg", 87913)),
-            "test_png_16rgba.png": (3435, 2090, ("png", 2090), ("webp", 728)),
-            "test_pnm.pnm": (27661, 27661, ("png", 15510), ("webp", 6458)),
+            "test_png_16rgba.png": (3435, 2090, ("png", 2090), ("webp", 1142)),
+            "test_pnm.pnm": (27661, 27661, ("png", 15510), ("webp", 12808)),
             "test_pre-optimized_png.png": (
                 256572,
                 256572,
                 ("png", 256572),
-                ("webp", 107674),
+                ("webp", 197726),
             ),
-            "eight.tif": (59640, 59640, ("png", 30564), ("webp", 18786)),
+            "eight.tif": (59640, 59640, ("png", 30564), ("webp", 24974)),
         }
     )
 else:
     FNS.update(
         {
-            "test_bmp.bmp": (141430, 141430, ("png", 67215), ("webp", 30530)),
+            "test_bmp.bmp": (141430, 141430, ("png", 67215), ("webp", 47524)),
             "test_gif.gif": (138952, 138944, ("png", 112137), ("webp", 107924)),
             "test_jpg.jpg": (97373, 87922, ("jpg", 87922), ("jpg", 87922)),
-            "test_png_16rgba.png": (3435, 2870, ("png", 2870), ("webp", 728)),
-            "test_pnm.pnm": (27661, 27661, ("png", 15510), ("webp", 6458)),
+            "test_png_16rgba.png": (3435, 2870, ("png", 2870), ("webp", 1142)),
+            "test_pnm.pnm": (27661, 27661, ("png", 15510), ("webp", 12808)),
             "test_pre-optimized_png.png": (
                 256572,
                 256572,
@@ -71,7 +71,7 @@ else:
 EXHAUSTIVE_FNS = MappingProxyType(
     {
         "test_png_16rgba.png": (3435, 2870, ("png", 2870), ("webp", 738)),
-        "test_webp_lossless.webp": (5334, 2044, ("webp", 2044), ("webp", 1946)),
+        "test_webp_lossless.webp": (5334, 3870, ("webp", 3870), ("webp", 1946)),
         "test_webp_lossless_pre-optimized.webp": (
             3798,
             3798,
@@ -141,13 +141,11 @@ class TestImagesDir(BaseTestImagesDir):
 class TestExhaustiveImageDir(BaseTestImagesDir):
     FNS = EXHAUSTIVE_FNS
 
-    def test_convert_to_webp_exhaustive(self) -> None:
+    def test_convert_to_webp_near_lossless(self) -> None:
         """Test convert to WEBP."""
         args = (
             PROGRAM_NAME,
-            "-rvvx",
-            "PNG",
-            "-ec",
+            "-rvvvnc",
             "WEBP",
             str(TMP_ROOT),
         )